# src/dtos/create-user.dto.ts
import { IsEmail, IsNotEmpty, MinLength } from 'class-validator';

export class CreateUserDto {
	@IsNotEmpty()
	name!: string;

	@IsEmail()
	email!: string;

	@MinLength(6)
	@IsNotEmpty()
	password!: string;
}